Moved mapping to main
- Test case to map has been moved to main.
- Renamed files.
