+++ Header/hybrid_dsatur.h
@@ -1,9 +1,9 @@
-#ifndef _MY_HYBRID_DSATUR_H_
-#define _MY_HYBRID_DSATUR_H_
+#ifndef _HYBRID_DSATUR_H_
+#define _HYBRID_DSATUR_H_
 
 #include "graph_color.h"
-#include "my_dsatur.h"
-#include "my_tabucol.h"
+#include "dsatur.h"
+#include "tabucol.h"
 
 using GraphColoring::Algorithm;
 using GraphColoring::GraphColor;
@@ -13,7 +13,9 @@ using GraphColoring::tabucol;
 namespace GraphColoring{
 	class hybrid_dsatur : public GraphColor {
 		public: 
-			hybrid_dsatur() :GraphColor() { algorithm = kHybridDSATUR; } 
+			hybrid_dsatur(map<string, vector<string> > input_graph) :GraphColor(input_graph) { 
+				algorithm = kHybridDSATUR; 
+			} 
 			map<string,int> color(int condition = 0);
 	};
 }
