+++ Source/dsatur.cpp
@@ -1,9 +1,6 @@
-#ifndef _DSATUR_H_
-#define _DSATUR_H_
+#include "../Header/dsatur.h"
 
-#include "../Header/graph.h"
-
-void GraphColoring::Graph::dsatur() {	
+map<string,int> GraphColoring::dsatur::color(int condition) {	
 
 	vector<string> todo;
 	string max_degree = "";
@@ -85,7 +82,8 @@ void GraphColoring::Graph::dsatur() {
 		if(saturation_name == "")
 		{
 			cerr << "Error: Could not find a max saturated node in the graph (reason unknown)" << endl;
-			return;
+			coloring.clear();
+			return coloring;
 		}
 
 		//We now know the most saturated node, so we remove it from the todo list
@@ -123,6 +121,6 @@ void GraphColoring::Graph::dsatur() {
 		}
 		saturation_level[saturation_name] = std::numeric_limits<int>::min();
 	}
+	return coloring;
 }
 
-#endif // _DSATUR_H_
