+++ Source/graph_color.cpp
@@ -12,16 +12,6 @@ using std::ifstream;
 using std::ofstream;
 using std::ostringstream;
 
-ifstream& Getline(ifstream& ifs,string& line) {
-    getline(ifs,line);
-    if (!line.empty()) {
-        while (isspace(line.at(line.size()-1))) {
-            line = line.substr(0,line.size()-1);
-        }
-    }
-    return ifs;
-}
-
 string GraphColoring::GraphColor::get_algorithm_string() { 
     switch (algorithm) {
         case kDSATUR:
@@ -53,153 +43,6 @@ map<string,int> GraphColoring::GraphColor::color(int condition) {
     return coloring;
 } 
 
-vector<string> GraphColoring::GraphColor::split(string to_split) {
-    vector<string> split_string;
-    unsigned index_start;
-    for (unsigned i = 0;i < to_split.length();i++) {
-        index_start = i;
-        while(i < to_split.length() && !isspace(to_split.at(i))) { i++; }
-        split_string.push_back(to_split.substr(index_start,i - index_start));
-    }
-    return split_string;
-}
-
-void GraphColoring::GraphColor::parse_edge_list(char* input_file) {
-    ifstream file(input_file);
-    if(file.is_open()) {
-        string line;
-        int vertices = -1;
-        int flag = 0;
-        while(!flag && Getline(file,line)) {
-            while(line.size() == 0) {
-                Getline(file,line);
-            }
-            vector<string> words = GraphColoring::GraphColor::split(line);
-            if(words.size() != 0) {
-                if(words[0] == "p") {
-                    vertices = atoi(words[2].c_str());
-                    flag = 1;
-                }
-            }
-        }
-        if(!flag || vertices == -3) {
-            cerr << "File is missing parameter line before edge list" << endl;
-            cerr << "Should be: \"p edge <number of vertices> <number of edges>\"" << endl;
-            return;
-        }
-        for(int i=0; i<vertices; i++) {
-            string pre = "v";
-            string temp;
-            std::ostringstream convert;
-            convert << (i+1);
-            temp = convert.str();
-            pre.append(temp);
-            vector<string> base;
-            graph[pre] = base;
-        }
-        while(Getline(file,line))
-        {
-            vector<string> words = GraphColoring::GraphColor::split(line);
-            if(words[0] == "e")
-            {
-                string arg1 = "v";
-                arg1.append(words[1]);
-                string arg2 = "v";
-                arg2.append(words[2]);
-                vector<string> base;
-                vector<string> base2;
-		graph[arg1].push_back(arg2);
-		graph[arg2].push_back(arg1);
-            }
-        }
-    } else {
-        cerr << "Input File Not Found" << endl;
-        return;
-    }
-}
-
-//Used to parse test inputs where the first line is the number of
-//vertices, and the next lines are the edge matrix
-void GraphColoring::GraphColor::parse_edge_matrix(char* input_file) {
-    string pre = "v";
-
-    ifstream file(input_file);
-    if(file.is_open()) {
-        string line;
-        Getline(file,line);
-        int n = atoi(line.c_str());
-        int i = 0;
-        while(Getline(file,line)) {
-            i += 1;
-            vector<string> words = GraphColoring::GraphColor::split(line);
-            if((int)words.size() != n) {
-                cerr << "Invalid Input, line " << i << " is not the correct length (" 
-                     << words.size() << "," << n << "): " << line << endl;
-                for (unsigned i = 0;i < words.size();i++) { cerr << "\t" << words.at(i) << endl; }
-                graph.clear();
-                return;
-            }
-            vector<string> edges;
-            for(int j = 0; j < n; j++) {
-                if(words[j] == "1") {
-                    pre = "v";
-                    string temp;
-                    ostringstream convert;
-                    convert << (j+1);
-                    temp = convert.str();
-                    edges.push_back(pre.append(temp));
-                }
-            }
-            pre = "v";
-            string temp;
-            ostringstream convert;
-            convert << i;
-            temp = convert.str();
-            graph[pre.append(temp)] = edges;
-        }
-        if(i != n) {
-            cerr << "Input is not the right length" << endl;
-            graph.clear();
-            return;
-        }
-        file.close();
-    } else {
-        cerr << "Input File Not Found" << endl;
-        return;
-    }
-    return;
-}
-
-vector< vector<string> > GraphColoring::GraphColor::get_input(char* input_file) {
-    vector< vector<string> > Input;
-
-    ifstream file(input_file);
-    if(file.is_open()) {
-        string line;
-        while(Getline(file,line)) {
-            vector<string> words = GraphColoring::GraphColor::split(line);
-            if(words.size() > 1) {
-                for(unsigned i=1; i<words.size(); i++) {
-                    if(words[i] != "0" && words[i] != "1" && words[i] != "x" && words[i] != "X") {
-                        cerr << "Problem with this Input Line: " << line << endl;
-                        cerr << "Problem is with word: \"" << words[i] << "\" at position " << i << endl;
-                        Input.clear();
-                        return Input;
-                    }
-                }
-            } else {
-                cerr << "No Actuation Data Found, Please Check Input File" << endl;
-                Input.clear();
-                return Input;
-            }
-            Input.push_back(words);
-        }
-        file.close();
-    } else {
-        cerr << "Input File Not Found" << endl;
-    }
-    return Input;
-}
 
 bool GraphColoring::GraphColor::verify() {
     for(map< string,vector<string> >::iterator i = graph.begin(); i != graph.end(); i++)
